CLI: Move non-bin files from bin/ to src/cli/
Follows-up https://github.com/qunitjs/qunit/pull/1271.
